A Script to process entire folders
This script processes all files within a directory, auto names them and uses the naming to check if a files has been previously processed, and if a face file has already been used on a specific video. 
As the script simply passes arguments to the run.py file, it is totally independent from the other code updates and should not cause any issues for merging.