Change the fps type from int to str
Change the fps type from int to str, for example: 2997/100, instead of 29

Using int type for fps is not accurate enough, causing slight differences in the length of the generated video compared to the original video, resulting in a slight deviation in the audio track. Changing fps to str type can solve this problem.